+++ CHANGELOG.md
@@ -11,7 +11,7 @@ The `Unreleased` section name is replaced by the expected version of next releas
 ### Added
 ### Changed
 
-- Extracted `Equinox.Codec` to external project `Gardelloyd`, with Json.net support in `Gardelloyd.NewtonsoftJson`
+- Extracted `Equinox.Codec` to external project `FsCodec`, with Json.net support in `FsCodec.NewtonsoftJson`
 
 ### Removed
 ### Fixed